backend/requirements.txt
google-genai>=1.33.0
pydantic>=2.9.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
slowapi>=0.1.9
pytest>=8.0.0
pytest-asyncio>=0.24.0
